refactor(login): refactor login app
Refs: TOCDEV-4477
